Make requests retry if they fail
Wrapping requests.get and requests.post functions to make sure that
failed requests are retried (this is cause the pnc instances are a bit
flaky) they currently retry 6 times. Right now all exceptions are caught
with the special case that ValueErrors are handled differently because
those are (as far as I'm concerned) caused by pnc returning bad json.